==> code/01-generate_ligand_library.ipynb <==
This notebook shows how to generate the virtual ligand library

import numpy as np
import itertools
from rdkit import Chem
import pandas as pd
# Si is used as placeholder
df = pd.read_csv("../ligand_pool/ligand_scaff_and_sub.csv")
df

scaffold_smi = df["scaffold"].dropna().to_list()[0]
R1_smi_lst = df["R1(P-sub)"].dropna().to_list()
R2_smi_lst = df["R2(C-sub)"].dropna().to_list()
R3_smi_lst = df["R3(N-sub)"].dropna().to_list()

scaffold = Chem.MolFromSmiles(scaffold_smi)
R1_lst = [Chem.MolFromSmiles(smi) for smi in R1_smi_lst]
R2_lst = [Chem.MolFromSmiles(smi) for smi in R2_smi_lst]
R3_lst = [Chem.MolFromSmiles(smi) for smi in R3_smi_lst]

R1_match = Chem.MolFromSmiles('P[SiH3]')
R2_match = Chem.MolFromSmiles('C[SiH3]')
R3_match = Chem.MolFromSmiles('N[SiH3]')
si_match = Chem.MolFromSmiles('[SiH4]')

R1_idx_lst = sorted([list(item) for item in scaffold.GetSubstructMatches(R1_match)],key=lambda x:x[1],reverse=True)
R2_idx_lst = sorted([list(item) for item in scaffold.GetSubstructMatches(R2_match)],key=lambda x:x[1],reverse=True)
R3_idx_lst = sorted([list(item) for item in scaffold.GetSubstructMatches(R3_match)],key=lambda x:x[1],reverse=True)

print(R1_idx_lst,R2_idx_lst,R3_idx_lst)
delete_idx = []
for item in R1_idx_lst:
    if item[0] > item[1]:
        item[0] -= 1
    delete_idx.append(item[1])


for idx in delete_idx:
    if idx < R2_idx_lst[0][0]:
        R2_idx_lst[0][0] -= 1
    if idx < R2_idx_lst[0][1]:
        R2_idx_lst[0][1] -= 1
if R2_idx_lst[0][0] > R2_idx_lst[0][1]:
    R2_idx_lst[0][0] -= 1
delete_idx.append(R2_idx_lst[0][1])
print(R1_idx_lst,R2_idx_lst,R3_idx_lst)

for idx in delete_idx:
    if idx < R3_idx_lst[0][0]:
        R3_idx_lst[0][0] -= 1
    if idx < R3_idx_lst[0][1]:
        R3_idx_lst[0][1] -= 1
if R3_idx_lst[0][0] > R3_idx_lst[0][1]:
    R3_idx_lst[0][0] -= 1
print(R1_idx_lst,R2_idx_lst,R3_idx_lst)

lig_smi_lst = []
group_lig_smi_map = {}
for item in itertools.product(R1_lst,R2_lst,R3_lst):
    R1,R2,R3 = item
    rw_mol = Chem.RWMol(scaffold)
    ## insert R1
    R1_atc_at_idx = R1_idx_lst[0][0]
    R2_atc_at_idx = R2_idx_lst[0][0]
    R3_atc_at_idx = R3_idx_lst[0][0]
    for idx_ in R1_idx_lst:
        rw_mol.RemoveAtom(idx_[1])

    si_idx_in_R1 = R1.GetSubstructMatch(si_match)[0]
    rw_R1 = Chem.RWMol(R1)
    nei_idx_to_si = []
    if rw_R1.GetNumAtoms() > 1:
        si_atom = rw_R1.GetAtomWithIdx(si_idx_in_R1)
        nei_to_si = si_atom.GetNeighbors()[0].GetIdx()
        
    rw_R1.RemoveAtom(si_idx_in_R1)
    nei_to_si = nei_to_si if si_idx_in_R1 > nei_to_si else nei_to_si - 1
    if rw_R1.GetNumAtoms() == 0:
        idx_0 = rw_mol.AddAtom(Chem.Atom(1))
        idx_1 = rw_mol.AddAtom(Chem.Atom(1))
        rw_mol.AddBond(R1_atc_at_idx,idx_0,Chem.BondType.SINGLE)
        rw_mol.AddBond(R1_atc_at_idx,idx_1,Chem.BondType.SINGLE)
    else:
        prev_at_num = rw_mol.GetNumAtoms()
        rw_mol.InsertMol(rw_R1)
        rw_mol.AddBond(R1_atc_at_idx,nei_to_si+prev_at_num,Chem.BondType.SINGLE)
        prev_at_num = rw_mol.GetNumAtoms()
        rw_mol.InsertMol(rw_R1)
        rw_mol.AddBond(R1_atc_at_idx,nei_to_si+prev_at_num,Chem.BondType.SINGLE)  
    ## insert R2
    for idx_ in R2_idx_lst:
        rw_mol.RemoveAtom(idx_[1])

    si_idx_in_R2 = R2.GetSubstructMatch(si_match)[0]
    rw_R2 = Chem.RWMol(R2)
    nei_idx_to_si = []
    if rw_R2.GetNumAtoms() > 1:
        si_atom = rw_R2.GetAtomWithIdx(si_idx_in_R2)
        nei_to_si = si_atom.GetNeighbors()[0].GetIdx()
        
    rw_R2.RemoveAtom(si_idx_in_R2)
    nei_to_si = nei_to_si if si_idx_in_R2 > nei_to_si else nei_to_si - 1
    if rw_R2.GetNumAtoms() == 0:
        idx_0 = rw_mol.AddAtom(Chem.Atom(1))
        rw_mol.AddBond(R2_atc_at_idx,idx_0,Chem.BondType.SINGLE)
    else:
        prev_at_num = rw_mol.GetNumAtoms()
        rw_mol.InsertMol(rw_R2)
        rw_mol.AddBond(R2_atc_at_idx,nei_to_si+prev_at_num,Chem.BondType.SINGLE)
    ## insert R3
    for idx_ in R3_idx_lst:
        rw_mol.RemoveAtom(idx_[1])
    si_idx_in_R3 = R3.GetSubstructMatch(si_match)[0]
    rw_R3 = Chem.RWMol(R3)
    nei_idx_to_si = []
    if rw_R3.GetNumAtoms() > 1:
        si_atom = rw_R3.GetAtomWithIdx(si_idx_in_R3)
        nei_to_si = si_atom.GetNeighbors()[0].GetIdx()
        
    rw_R3.RemoveAtom(si_idx_in_R3)
    nei_to_si = nei_to_si if si_idx_in_R3 > nei_to_si else nei_to_si - 1
    if rw_R3.GetNumAtoms() == 0:
        idx_0 = rw_mol.AddAtom(Chem.Atom(1))
        rw_mol.AddBond(R3_atc_at_idx,idx_0,Chem.BondType.SINGLE)
    else:
        prev_at_num = rw_mol.GetNumAtoms()
        rw_mol.InsertMol(rw_R3)
        rw_mol.AddBond(R3_atc_at_idx,nei_to_si+prev_at_num,Chem.BondType.SINGLE)

    lig_mol = rw_mol.GetMol()
    lig_smi = Chem.MolToSmiles(lig_mol)
    lig_smi_lst.append(lig_smi)
    group_lig_smi_map[(Chem.MolToSmiles(R1),Chem.MolToSmiles(R2),Chem.MolToSmiles(R3))] = lig_smi
print(f"The size of ligand pool: {len(group_lig_smi_map)}")

np.save('../ligand_pool/vir_lig_smi_map.npy',group_lig_smi_map)
