==> code/02-generate_complex_graph.ipynb <==
This notebook shows how to generate the intermediate complexes.

import pandas as pd
import numpy as np
from tqdm import tqdm
from rdkit import Chem
from asymopt.geom import build_metal_sub_lig_comp_mol # To use the function, please run "pip install ." in the parent directory

lit_data = pd.read_csv("../data/processed/lit_data.csv")
lit_rct_1_smi_lst = lit_data["rct_1_smi"].to_list()
lit_rct_2_smi_lst = lit_data["rct_2_smi"].to_list()
lit_metal_lst = lit_data["metal"].to_list()
lit_lig_smi_lst = lit_data["lig_smi"].to_list()

## generate complex graph
lit_complex_lst = [build_metal_sub_lig_comp_mol(rct_1,rct_2,metal,lig) for rct_1, rct_2, metal, lig in zip(lit_rct_1_smi_lst, 
                                                                                                           lit_rct_2_smi_lst, 
                                                                                                           lit_metal_lst, 
                                                                                                           lit_lig_smi_lst)]
lit_complex_smi_lst = [Chem.MolToSmiles(mol) for mol in tqdm(lit_complex_lst)]

lit_data["complex_smi"] = lit_complex_smi_lst
lit_data.to_csv("../data/processed/lit_data_with_cplex_smi.csv", index=False)
