==> code/03-generate_2d_descriptor.ipynb <==
from asymopt.desc import getmorganfp,getPhysChemDesc
import pandas as pd
from rdkit import Chem
from tqdm import tqdm
import numpy as np

lit_data = pd.read_csv("../data/processed/lit_data_with_cplex_smi.csv")
lit_complex_smi_lst = lit_data["complex_smi"].to_list()
lit_complex_lst = [Chem.MolFromSmiles(smi) for smi in lit_complex_smi_lst]

## Generate 2D descriptor
lit_complex_fp_lst = []
lit_complex_rdkit_lst = []
for mol in tqdm(lit_complex_lst):
    Chem.SanitizeMol(mol)
    fp = getmorganfp(mol)
    rdkit_desc = getPhysChemDesc(mol)
    lit_complex_fp_lst.append(fp)
    lit_complex_rdkit_lst.append(np.array(rdkit_desc))
lit_complex_fp_arr = np.array(lit_complex_fp_lst)
lit_complex_rdkit_arr = np.array(lit_complex_rdkit_lst)

lit_complex_fp_arr[:5], lit_complex_rdkit_arr[:5]
