==> code/04-calculate_sascore.ipynb <==
import numpy as np
from rdkit import Chem
from rdkit.Contrib.SA_Score.sascorer import calculateScore
from rdkit import RDLogger
RDLogger.DisableLog('rdApp.*')

vir_lig_smi_map = np.load("../ligand_pool/vir_lig_smi_map.npy", allow_pickle=True).item()
vir_lig_smi_lst = list(vir_lig_smi_map.values())
smi_sascore_map = {smi: calculateScore(Chem.MolFromSmiles(smi)) for smi in vir_lig_smi_lst}
np.save("../ligand_pool/smi_sascore_map.npy",smi_sascore_map)

smi_sascore_map
